Made rownr into PRIMARY KEY, speeds up some queries; added debug code for query time measurement

+++ webapp/src/main/java/org/entrystore/rowstore/store/Dataset.java
@@ -76,7 +76,7 @@ public interface Dataset {
 	/**
 	 * @return Returns the size (amount of rows) of the dataset.
 	 */
-	int getRowCount();
+	long getRowCount();
 
 	/**
 	 * Returns all aliases of the dataset.
